add in test and nits
