+++ wandb/sdk/artifacts/_validators.py
@@ -3,10 +3,16 @@
 from __future__ import annotations
 
 import re
-from dataclasses import dataclass, field
 from functools import wraps
-from typing import TYPE_CHECKING, Any, Callable, TypeVar, cast, overload
+from typing import TYPE_CHECKING, Any, Callable, Literal, TypeVar, cast, overload
 
+from pydantic import Field
+from pydantic.dataclasses import dataclass
+
+from wandb.automations._generated.fragments import (
+    ArtifactPortfolioScopeFields,
+    ArtifactSequenceScopeFields,
+)
 from wandb.sdk.artifacts.exceptions import (
     ArtifactFinalizedError,
     ArtifactNotLoggedError,
@@ -28,8 +34,12 @@ MAX_ARTIFACT_METADATA_KEYS: Final[int] = 100
 ARTIFACT_NAME_MAXLEN: Final[int] = 128
 ARTIFACT_NAME_INVALID_CHARS: Final[frozenset[str]] = frozenset({"/"})
 
-LINKED_ARTIFACT_COLLECTION_TYPE: Final[str] = "ArtifactPortfolio"
-SOURCE_ARTIFACT_COLLECTION_TYPE: Final[str] = "ArtifactSequence"
+LINKED_ARTIFACT_COLLECTION_TYPE: Final[str] = ArtifactPortfolioScopeFields.model_fields[
+    "typename__"
+].default
+SOURCE_ARTIFACT_COLLECTION_TYPE: Final[str] = ArtifactSequenceScopeFields.model_fields[
+    "typename__"
+].default
 
 
 @dataclass
@@ -39,20 +49,13 @@ class _LinkArtifactFields:
     entity_name: str
     project_name: str
     name: str
-    version_index: int
+    version: str
     aliases: list[str]
 
-    # These fields shouldn't be set by the user as they should always be these values for a linked artifact
-    _is_link: bool = field(init=False, default=False)
-    _linked_artifacts: list[Artifact] = field(init=False, default_factory=list)
-
-    @property
-    def is_link(self) -> bool:
-        return self._is_link
-
-    @property
-    def linked_artifacts(self) -> list[Artifact]:
-        return self._linked_artifacts
+    # These fields shouldn't be set as they should always be
+    # these values for a linked artifact
+    is_link: Literal[False] = Field(init=False, frozen=True, default=True)
+    linked_artifacts: list = Field(init=False, frozen=True, default_factory=list)
 
 
 # For mypy checks
